+++ src/plugins/fontawesome.js
@@ -3,8 +3,9 @@ import Vue from 'vue';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 
-import { faHouse } from '@fortawesome/free-solid-svg-icons';
+import { faAngleDown, faAngleUp, faClock, faEnvelope } from '@fortawesome/free-solid-svg-icons';
+import { faWeixin, faQq, faWeibo } from '@fortawesome/free-brands-svg-icons';
 
-library.add([faHouse]);
+library.add([faAngleDown, faAngleUp, faClock, faEnvelope, faWeixin, faQq, faWeibo]);
 
 Vue.component('font-awesome-icon', FontAwesomeIcon);